Moved functionality to data module.
