+++ stellargraph/data/edge_data.py
